feat: finish user model
