--- app/src/main/java/dog/snow/androidrecruittest/data/model/RawUser.kt
@@ -1,36 +1,0 @@
-package dog.snow.androidrecruittest.data.model
-
-import android.os.Parcelable
-import dog.snow.androidrecruittest.data.model.util.Id
-import kotlinx.android.parcel.Parcelize
-
-@Parcelize
-data class RawUser(
-    val id: Id,
-    val name: String,
-    val username: String,
-    val email: String,
-    val address: RawAddress,
-    val phone: String,
-    val website: String,
-    val company: RawCompany
-) : Parcelable {
-    @Parcelize
-    data class RawAddress(
-        val street: String,
-        val suite: String,
-        val city: String,
-        val zipcode: String,
-        val geo: RawGeo
-    ) : Parcelable {
-        @Parcelize
-        data class RawGeo(val lat: String, val lng: String) : Parcelable
-    }
-
-    @Parcelize
-    data class RawCompany(
-        val name: String,
-        val catchPhrase: String,
-        val bs: String
-    ) : Parcelable
-}
