+++ app/src/main/java/dog/snow/androidrecruittest/data/model/common/Id.kt
@@ -1,4 +1,4 @@
-package dog.snow.androidrecruittest.data.model.util
+package dog.snow.androidrecruittest.data.model.common
 
 import android.os.Parcelable
 import kotlinx.android.parcel.Parcelize
