+++ app/src/main/java/dog/snow/androidrecruittest/data/model/common/Name.kt
@@ -1,0 +1,7 @@
+package dog.snow.androidrecruittest.data.model.common
+
+import android.os.Parcelable
+import kotlinx.android.parcel.Parcelize
+
+@Parcelize
+data class Name(val value: String) : Parcelable
