+++ app/src/main/java/dog/snow/androidrecruittest/data/model/geolocation/Coordinate.kt
@@ -1,0 +1,7 @@
+package dog.snow.androidrecruittest.data.model.geolocation
+
+import android.os.Parcelable
+import kotlinx.android.parcel.Parcelize
+
+@Parcelize
+data class Coordinate(val value: String) : Parcelable
