+++ app/src/main/java/dog/snow/androidrecruittest/data/model/type/City.kt
@@ -1,0 +1,7 @@
+package dog.snow.androidrecruittest.data.model.type
+
+import android.os.Parcelable
+import kotlinx.android.parcel.Parcelize
+
+@Parcelize
+data class City(val value: String) : Parcelable
