+++ app/src/main/java/dog/snow/androidrecruittest/data/model/type/ZipCode.kt
@@ -1,0 +1,7 @@
+package dog.snow.androidrecruittest.data.model.type
+
+import android.os.Parcelable
+import kotlinx.android.parcel.Parcelize
+
+@Parcelize
+data class ZipCode(val value: String) : Parcelable
