+++ app/src/main/java/dog/snow/androidrecruittest/data/model/user/Email.kt
@@ -1,0 +1,7 @@
+package dog.snow.androidrecruittest.data.model.user
+
+import android.os.Parcelable
+import kotlinx.android.parcel.Parcelize
+
+@Parcelize
+data class Email(val value: String) : Parcelable
