+++ app/src/main/java/dog/snow/androidrecruittest/data/model/user/UserName.kt
@@ -1,0 +1,7 @@
+package dog.snow.androidrecruittest.data.model.user
+
+import android.os.Parcelable
+import kotlinx.android.parcel.Parcelize
+
+@Parcelize
+data class UserName(val value: String) : Parcelable
