+++ app/src/main/java/dog/snow/androidrecruittest/data/model/user/Website.kt
@@ -1,0 +1,7 @@
+package dog.snow.androidrecruittest.data.model.user
+
+import android.os.Parcelable
+import kotlinx.android.parcel.Parcelize
+
+@Parcelize
+data class Website(val value: String) : Parcelable
